+++ inst/couple-Atlantis/cpolar_coupling_z_index.R
@@ -40,6 +40,7 @@ library(rbgm) ## read BGM
 library(bgmfiles) ## archive of BGM files
 library(raadtools)
 library(ncdump)
+library(dplyr)
 ## get the Circumpolar files
 cpolar <- raadtools:::cpolarfiles()
 
@@ -69,7 +70,8 @@ roms_face <- romsmap(project_to(faceSpatial(bgm), "+init=epsg:4326"), roms_ll)
 ## which box does each point fall in
 index_box <- function(box_sp, roms_ll) {
   ind <- sp::over(project_to(coords_points(roms_ll), box_sp) , as(box_sp, "SpatialPolygons"))
-  tibble(box = ind, 
+#  tibble(box = ind,
+  tibble(box = box_sp$label[ind], 
          cell = seq_len(ncell(roms_ll))) %>% 
     filter(!is.na(box))
 }
@@ -80,19 +82,16 @@ index_box <- function(box_sp, roms_ll) {
 library(dplyr)
 box_roms_index <- index_box(boxes, roms_ll)
 ind_face <- cellFromLine(romsdata(roms_file, "u"), roms_face)
-face_roms_index <- tibble(face = rep(seq_len(nrow(roms_face)), lengths(ind_face)), 
+face_roms_index <- tibble(face = roms_face$label[rep(seq_len(nrow(roms_face)), lengths(ind_face))], 
                           cell = unlist(ind_face))
 
 
 #' return the ramp of positive depths from surface down 
 #' (so that the order is native to the NetCDF order)
 roms_level <- function(Cs_r, h, cell) {
-  extract(h, cell) *  -rev(Cs_r)
+  extract(h, cell) *  Cs_r
 }
-## put any raster into xy-index space (0, nc, 0, nr)
-set_indextent <- function(x) {
-  setExtent(x, extent(0, ncol(x), 0, nrow(x)))
-}
+
 ## important to readAll here, else extract is very slow in the loop
 h <- readAll(raster(roms_file, varname = "h"))
 ## Cs_r is the S-coord stretching
@@ -103,24 +102,41 @@ Cs_r <- rawdata(roms_file, "Cs_r")
 
 list_nc_z_index <- vector('list', nrow(box_roms_index))
 max_depth <- max(extract(h, unique(box_roms_index$cell)))
-atlantis_depths <- cumsum(c(0, rev(rbgm::build_dz(-max_depth))))
+atlantis_depths <- -rev(cumsum(c(0, rev(rbgm::build_dz(-max_depth)))))
 for (i in seq_len(nrow(box_roms_index))) {
   rl <- roms_level(Cs_r, h, box_roms_index$cell[i])
   ## implicit 0 at the surface, and implicit bottom based on ROMS
-  list_nc_z_index[[i]] <- findInterval(rl, atlantis_depths)
+  list_nc_z_index[[i]] <- length(atlantis_depths) - findInterval(rl, atlantis_depths) + 1
 if (i %% 1000 == 0) print(i)
 }
+# 
+# ll <- extract(roms_ll, box_roms_index$cell)
+# box_roms_index$lon <- ll[, 1]
+# box_roms_index$lat <- ll[, 2]
+# library(ggplot2)
+# ggplot(box_roms_index, aes(lon, lat, colour = box)) + geom_point(pch = ".")
+# 
+
 ## join the box-xy-index to the level index
 box_z_index <- bind_rows(lapply(list_nc_z_index, 
-                 function(x) tibble(atlantis_level = x, roms_level = seq_along(x))), 
+                 function(x) tibble(atlantis_level = pmax(1, x), roms_level = seq_along(x))), 
           .id = "cell_index") %>% 
   inner_join(mutate(box_roms_index, cell_index = as.character(row_number()))) %>% 
   select(-cell_index)
 
+
+# ll <- extract(roms_ll, box_z_index$cell)
+# box_z_index$lon <- ll[, 1]
+# box_z_index$lat <- ll[, 2]
+# library(ggplot2)
+# ggplot(box_z_index, aes(lon, lat, colour = roms_level)) + geom_point(pch = ".") + 
+#   facet_wrap(~atlantis_level)
+
+
 for (i in seq_len(nrow(face_roms_index))) {
   rl <- roms_level(Cs_r, h, face_roms_index$cell[i])
   ## implicit 0 at the surface, and implicit bottom based on ROMS
-  list_nc_z_index[[i]] <- findInterval(rl, atlantis_depths)
+  list_nc_z_index[[i]] <- length(atlantis_depths) -findInterval(rl, atlantis_depths) + 1
   if (i %% 1000 == 0) print(i)
 }
 ## join the face-xy-index to the level index
@@ -141,14 +157,16 @@ extract_at_level <- function(x, cell_level) {
   values <- numeric(nrow(cell_level))
   for (ul in seq_along(ulevel)) {
     asub <- cell_level$level == ulevel[ul]
-    values[asub] <- as.vector(extract(brick(x[[ulevel[ul]]])), 
+    values[asub] <- extract(x[[ulevel[ul]]], 
             cell_level$cell[asub])
   }
   values
 }
 
+#file_db <- file_db[1, ]
 box_props <- face_props <- vector("list", nrow(file_db))
-for (i_timeslice in seq(nrow(file_db))) {
+i_timeslice <- 1
+#for (i_timeslice in seq(nrow(file_db))) {
   roms_file <- file_db$fullname[i_timeslice]
   level <- file_db$band_level[i_timeslice]
   ru <- set_indextent(brick(roms_file, varname = "u", lvar = 4, level = level))
@@ -165,33 +183,37 @@ for (i_timeslice in seq(nrow(file_db))) {
   face_props[[i_timeslice]] <-  face_z_index %>% group_by(atlantis_level, face) %>% 
     summarize(flux = mean(sqrt(u * u + v * v), na.rm = TRUE) * 24 * 3600) %>% 
     mutate(band_level = level)
-}
+#}
 
+box_props <- bind_rows(box_props)
+face_props <- bind_rows(face_props)
+library(ggplot2)
+ggplot(box_props, aes(x = temp, y = salt, colour = atlantis_level)) + geom_point() + facet_wrap(~box)
+ggplot(face_props, aes(x = atlantis_level, y = flux, colour = atlantis_level)) + geom_point() 
 
-
-
-
-plot(romsmap(boxes[15, ], roms_ll), add = F, border = "red", lwd = 4)
-plot(sqrt(ru[[10]]^2 + rv[[10]]^2), add = TRUE, 
-     col = colorRampPalette(c("#132B43", high = "#56B1F7"))(10), zlim = c(0, 0.1))
-plot(romsmap(boxes[15, ], roms_ll), add = TRUE, border = "red", lwd = 4)
-ggplot(mutate(z_index %>% 
-                filter(box == 15), 
-              x = xFromCell(ru, cell), y = yFromCell(ru, cell)), 
-       aes(x = x, y = y, fill = sqrt(u * u + v * v))) + geom_tile() + coord_equal() + 
-  facet_wrap(~roms_level)
-
-
-rm(ru, rv)
-
-
-ggplot(z_index %>% mutate(x = xFromCell(rdata, cell), 
-                          y = yFromCell(rdata, cell)), 
-       aes(x = x, y = y, fill = temp)) + geom_tile() + 
-  coord_equal()
-
-ggplot(mutate(z_index %>% 
-                filter(roms_level %% 3 == 0), 
-              x = xFromCell(ru, cell), y = yFromCell(ru, cell)), 
-       aes(x = x, y = y, fill = temp)) + geom_tile() + coord_equal() + 
-  facet_wrap(~roms_level)
+# 
+# 
+# plot(romsmap(boxes[15, ], roms_ll), add = F, border = "red", lwd = 4)
+# plot(sqrt(ru[[10]]^2 + rv[[10]]^2), add = TRUE, 
+#      col = colorRampPalette(c("#132B43", high = "#56B1F7"))(10), zlim = c(0, 0.1))
+# plot(romsmap(boxes[15, ], roms_ll), add = TRUE, border = "red", lwd = 4)
+# ggplot(mutate(z_index %>% 
+#                 filter(box == 15), 
+#               x = xFromCell(ru, cell), y = yFromCell(ru, cell)), 
+#        aes(x = x, y = y, fill = sqrt(u * u + v * v))) + geom_tile() + coord_equal() + 
+#   facet_wrap(~roms_level)
+# 
+# 
+# rm(ru, rv)
+# 
+# 
+# ggplot(z_index %>% mutate(x = xFromCell(rdata, cell), 
+#                           y = yFromCell(rdata, cell)), 
+#        aes(x = x, y = y, fill = temp)) + geom_tile() + 
+#   coord_equal()
+# 
+# ggplot(mutate(z_index %>% 
+#                 filter(roms_level %% 3 == 0), 
+#               x = xFromCell(ru, cell), y = yFromCell(ru, cell)), 
+#        aes(x = x, y = y, fill = temp)) + geom_tile() + coord_equal() + 
+#   facet_wrap(~roms_level)